Combine notify_abort into notify plugin
